Try to restore quick-fix windows...
The code seems simple enough and in fact works most of the time. The
problem is that it doesn't always work. I've added it anyway because
of its added value to my (and presumably others) work flow. If and when
I find a more reliable way to restore quick-fix windows I can always
switch to that.