Remember the 'background' setting
I love the Earendel color scheme, amongst other things because it looks
great on light as well as dark backgrounds. During the day I use a light
background color and in the evening/night I switch to a dark background
color so as not to strain my eyes. What always really annoyed me is that
the session plug-in wouldn't remember this setting. Now it will! :-)