Ignore missing fonts (reported by SteP)
When the session scripts are stored on a removable drive which is used
on multiple computers the session scripts can reference a font which
existed on the previous machine but not the current one, causing an
E596 error to trigger.

SteP also reported that on Windows XP SP3 after the E596 error triggers,
Vim becomes unresponsive. I can't reproduce this on my virtual machine
running the same version of Windows but I'm hoping this change will also
fix that problem (it's my understanding that :silent! and :try/:catch
are equivalent in their ability to silence errors).