Bug fix: Don't hang during :RestartVim when &verbose >= 1
When I set &verbose >= 1 and use :RestartVim it sometimes hangs just
before restarting with the dreaded "hit enter" prompt. I guess this
happens because Vim prints "Writing viminfo file" and the message
is about to scroll by without the user noticing. Anyway shutting down
the old instance of Vim is half the point of the :RestartVim command
so now it silences the "Writing viminfo file" message and others!