ENH: Ask to skip restoring for now, or delete the last-session.txt.
Change the check for restoring a previously persisted session so that there's a distinction between
"No": Then, Vim will not ask for restoring the session on subsequent startups. This is implemented by deleting the last-session.txt file.
"Skip": Do not restore the session now, but ask again on the next Vim startup (the old behavior).

Also, introduce a g:session_verbose_messages configuration variable; I found the hint how to disable the dialog distracting, because the long explanation makes it hard to focus on the actual decision.