No longer try to track dirty sessions; assume always dirty.
Trying to track dirty sessions was causing more problems than it solved. It only took into account the list of buffers and window layout, ignoring all the other options than get saved with a session that may have changed. There's no good way to know what options are getting saved with the session, so for now, assume a session is always dirty.

This will cause you to always be prompted to save a session upon closing a session, but that's the trade-off.