FIX: Determine tab / global scope from sourced session, not the config.
For the message after opening a session, use the t:this_session / v:this_session information from the loaded session instead of the current configuration. This matters especially when a session is auto-loaded on Vim startup, because then the scope cannot be determined from the command, (since there is no command).
The :OpenSession / :OpenTabSession commands actually perform the same (maybe there should be a warning when a global session is opened through :OpenTabSession?!)