Remove the -count=94919 workaround for :AppendTabSession.
As the command only takes the optional count _before_ the command (like :tabnew, as per the documentation), no count can be detected by -1 if we specify -range=-1.
This also avoids the unexpected behavior of -count that a passed argument in the form "42foo" is parsed as count=42 and args=foo. With this change, one can properly define session names that start with numbers.