Handle (i.e. do not increase) zero-width windows.
The :windo command enters each window, which causes the height to increase to at least 1 (to make the current line visible). (The so-called "Rolodex mode" uses stacks of 0-width windows, with only one maximized.)
To avoid changes to the window layout, it (and the current window number) need to be saved and restored after the :windo. This affects the iteration in s:check_special_tabpage() (during :SessionSave) as well as the commands added to the session (on :SessionOpen restoring a single tab page).