Write correct session lock when (auto-)opening tab-scoped session.
When a tab-scoped session is automatically opened during load (or via the :OpenSession command), the session lock does not contain the 'tabpage' attribute. This causes the auto-save when quitting Vim to fail with "The 'default' session is already loaded in this Vim. Use the command ':SaveTabSession! default' to override."

During auto-load (because we cannot know whether the to-be-loaded session is a session-scoped one), we need to update the lock in case of a tab-scoped session _after_ opening the session.
Additionally, the s:vim_instance_id() function also needs to consider xolox#session#is_tab_scoped(), as only that is set during (auto-)load; xolox#session#include_tabs() only gives the correct result when one of the :...TabSession commands is invoked.