Update screenshots of automatic session save/restore in README
I had already planned to update the screenshots after merging pull
request #104 but Rory was quicker than me by creating issue #117 :-).
Silly as it sounds it took me a while to take somewhat representative
screenshots because I've been using a tiling window manager (Awesome)
for the past couple of years which hides all window decorations ...

One thing to note about this change: I used full URLs to the screenshots
by choice (even though I know that in general this is not the
recommended way) because README.md is used in more places than only
GitHub (specifically I also use it to render the homepage on my website
and I convert it to a Vim help file to be bundled with the Vim plug-in).

See also issue 104 on GitHub:
  https://github.com/xolox/vim-session/pull/104

See also issue 117 on GitHub:
  https://github.com/xolox/vim-session/issues/117